requirements.txt
pytest-playwright==0.7.2
python-dotenv==1.2.2
Faker==40.8.0